Modified Tile selection
Modified tile selection in cScene for the isometric coordinates labeling

+++ cScene.h
@@ -41,7 +41,7 @@ public:
 	void Move(int pointer); // ajustar condicions? --v
 	void MoveByRadar(int x,int y);
 	bool Visible(int cellx,int celly);
-	void TileSelected(int mouse_x,int mouse_y,int *tile_x,int *tile_y,int *atx,int *aty,int *Dx,int *Dy);
+	void TileSelected(int mouse_x,int mouse_y,float *tile_x, float *tile_y,float *atx,float *aty,float *Dx,float *Dy);
 
 	void IsoToCartesian(float ix, float iy, int& cx, int& cy); // Method to tansform isometric to cartesian coordinates
 	void CartesianToIso(int cx, int cy, float& ix, float& iy); // Method to tansform cartesian to isometric coordinates
